Add support for strict npm modules.
If you import elasticlunr, and the parser has strict
settings enabled, the parser complains that lunr is not
defined.

This change creates a version that defines lunr as a var
instead of a global, which fixes this error. This makes it
so elasticlunr is no longer a drop in replace for lunr, but
when using npm, it seems that a developer would expect to
do something like:

  `import lunr from 'elasticlunr';`

or

  `lunr = require('elasticlunr');`